Converting the project to a windows app, displaying my monsters and having them move randomly.

+++ Engine/Point2D.cpp
@@ -8,17 +8,17 @@ Point2D::Point2D()
 	y = 0;
 }
 
-Point2D::Point2D(int i_x, int i_y)
+Point2D::Point2D(float i_x, float i_y)
 {
 	x = i_x;
 	y = i_y;
 }
 
-int Point2D::getX() { return x; }
-int Point2D::getY() { return y; }
+float Point2D::getX() { return x; }
+float Point2D::getY() { return y; }
 
-void Point2D::setX(int i_x) { x = i_x; }
-void Point2D::setY(int i_y) { y = i_y; }
+void Point2D::setX(float i_x) { x = i_x; }
+void Point2D::setY(float i_y) { y = i_y; }
 
 void Point2D::print()
 {
@@ -53,14 +53,14 @@ Point2D Point2D::operator/=(Point2D other)
 	return Point2D(x, y);
 }
 
-Point2D Point2D::operator*=(int other)
+Point2D Point2D::operator*=(float other)
 {
 	x *= other;
 	y *= other;
 	return Point2D(x, y);
 }
 
-Point2D Point2D::operator/=(int other)
+Point2D Point2D::operator/=(float other)
 {
 	x /= other;
 	y /= other;
@@ -105,22 +105,22 @@ Point2D operator/(Point2D lhs, Point2D rhs)
 	return Point2D(lhs.getX() / rhs.getX(), lhs.getY() / rhs.getY());
 }
 
-Point2D operator*(Point2D lhs, int rhs)
+Point2D operator*(Point2D lhs, float rhs)
 {
 	return Point2D(lhs.getX() * rhs, lhs.getY() * rhs);
 }
 
-Point2D operator*(int lhs, Point2D rhs)
+Point2D operator*(float lhs, Point2D rhs)
 {
 	return Point2D(lhs * rhs.getX(), lhs * rhs.getY());
 }
 
-Point2D operator/(Point2D lhs, int rhs)
+Point2D operator/(Point2D lhs, float rhs)
 {
 	return Point2D(lhs.getX() / rhs, lhs.getY() / rhs);
 }
 
-const Point2D Point2D::Up = Point2D(0, 1);
-const Point2D Point2D::Down = Point2D(0, -1);
-const Point2D Point2D::Left = Point2D(-1, 0);
-const Point2D Point2D::Right = Point2D(1, 0);
+const Point2D Point2D::Up = Point2D(0.0, 1.0);
+const Point2D Point2D::Down = Point2D(0.0, -1.0);
+const Point2D Point2D::Left = Point2D(-1.0, 0.0);
+const Point2D Point2D::Right = Point2D(1.0, 0.0);
